Complete Sample Data
I finished manually entering the class schedules into the xlsx spreadsheet and duplicate this spreadsheet as a CSV. This then allowed me to add functionality to the Java Master Class to parse the CSV for the list of tutors and their availabilities. The Java code can now read this CSV file and compile for each tutor a list of their availabilities as well as create a HashMap mapping Shifts to a list of Tutors available during that shift. The Java also only adds shifts to a tutor's availability if that tutor is available for 2 or more hours in a row.