schedule generates according to inputted 2D Array
There's a lot to this commit as I haven't committed in a while. Main things being a lot of changes to get the php of the weeklySchedule and the Javascript of the forms to communicate sufficiently. Got the php part working with a tester.js file, so I know the php is working. Now I need to find a way to get the Javascript to save the schedule 2D array so that it can be properly passed to the php.