updated scheduling algorithm to consider more constraints
Added veteranStatus variable to Tutor object (not yet used). Changed sample data to include subjects (i.e. the tutors' roles at the center). Chenged Master's readFile() method to read and process the updated sample data to store the subject information within tutor objects (in a newly created HashSet<String> attribute for Tutor objects). Changed Master's refreshScores() method to reheapify the shiftQ (this was a logic bug in the behavior of the scheduling algorithm). Changed the scheduling algorithm's behavior in multiple places to .remove() a tutor from a shift's availableTutor queue if we remove that tutor from their assignedTutors list (or if the algorithm decided it can't assign the tutor to that shift when trying to assign the tutor to that shift during the assignConsecutiveShift method); before, the algorithm would keep the tutor available for that shift, but this often led to infinite loops and this seemed like the easiest solution -- if I have time, I'd like to come back and just have the algorithm increase the tutor's score for that shift so that they don't come up again before someone else has had a chance. Implemented behavior ensuring that all shifts have at least one writing advisor. Adjusted Shift's hasMin() method to allow 9am shifts to be worked by only 1 person. Adjusted Shift's adjustScore() method to hyperbolically adjust scores with more tutors (rather than linearly).